openApiUtils put back in its own file (also used by GraphQL)
